@@ -10,11 +10,11 @@ namespace ChaosSoft.Core.Data
             Y = y;
         }
 
-        public double X { get; set; }
+        public double X { get; }
 
-        public double Y { get; set; }
+        public double Y { get; }
 
         public override string ToString() =>
-            $"{NumFormat.ToShort(X)}, {NumFormat.ToShort(Y)}";
+            $"{NumFormatter.ToShort(X)}, {NumFormatter.ToShort(Y)}";
     }
 }
