@@ -1,8 +1,10 @@
 ﻿using ChaosSoft.Core.Data;
+using ChaosSoft.Core.Extensions;
 using ChaosSoft.Core.IO;
-using ChaosSoft.Core.NumericalMethods.EmbeddingDimension;
+using ChaosSoft.Core.NumericalMethods.PhaseSpace;
 using System;
 using System.Diagnostics;
+using System.Linq;
 using System.Text;
 
 namespace ChaosSoft.Core.NumericalMethods.Lyapunov
@@ -10,93 +12,91 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
     /// <summary>
     /// M.Sano and Y.Sawada, Measurement of the Lyapunov spectrum from a chaotic time series, Phys. Rev.Lett. 55, 1082 (1985).
     /// </summary>
-    public class LeSpecSanoSawada : LyapunovMethod
+    public sealed class LeSpecSanoSawada : ITimeSeriesLyapunov, IDescribable
     {
         private const string Paper = "M.Sano and Y.Sawada, Measurement of the Lyapunov spectrum from a chaotic time series, Phys. Rev.Lett. 55, 1082 (1985)";
         private const int OutputInterval = 10000;
         private const double EpsMax = 1.0;
 
-        private readonly BoxAssistedFnn fnn;
-
         private readonly int _eDim;
         private readonly bool _inverse;
-        private readonly int _minNeighbors = 30;
-        private readonly double _epsStep = 1.2;
-        private readonly int _iterations;
-        private readonly int _length;
+        private readonly int _minNeighbors;
+        private readonly bool _epsSet;
+        private readonly double _epsStep;
         private readonly int _tau;
+        private int iterations;
 
-        private readonly bool _epsset = false;
-
-        private double averr;
-        private double avneig = 0.0, aveps = 0.0;
+        private double errorAvg;
+        private double neighborsAvg = 0;
+        private double epsAvg = 0;
+        private double epsMin;
         private double[,] matrix;
-        private double[] vector, abstand;
-        private double epsmin;
+        private double[] vector;
+        private double[] abstand;
         private long count = 0;
         private int[] indexes;
-        private Random random;
 
         /// <summary>
         /// The method estimates the whole spectrum of Lyapunov exponents for a given, possibly multivariate, time series. 
         /// Whole spectrum means: If d components are given and the embedding dimension is m than m*d exponents will be determined. 
         /// The method is based on the work of Sano and Sawada.
         /// </summary>
-        /// <param name="timeSeries"></param>
         /// <param name="eDim"></param>
+        /// <param name="tau"></param>
         /// <param name="iterations"></param>
         /// <param name="scaleMin"></param>
         /// <param name="epsstep"></param>
         /// <param name="minNeigh"></param>
         /// <param name="inverse"></param>
-        public LeSpecSanoSawada(double[] timeSeries, int eDim, int tau, int iterations, double scaleMin, double epsstep, int minNeigh, bool inverse)
-            : base(timeSeries)
+        public LeSpecSanoSawada(int eDim, int tau, int iterations, double scaleMin, double epsstep, int minNeigh, bool inverse)
         {
             _eDim = eDim;
             _tau = tau;
-            _iterations = iterations;
-            epsmin = scaleMin;
-            _epsset = scaleMin != 0;
+            this.iterations = iterations;
+            epsMin = scaleMin;
+            _epsSet = scaleMin != 0;
             _epsStep = epsstep;
             _minNeighbors = minNeigh;
             _inverse = inverse;
-
-            _length = Series.Length;
-
-            if (_minNeighbors > (_length - _tau * (eDim - 1) - 1))
-            {
-                throw new ArgumentException($"Too few points to find {_minNeighbors} neighbors, it makes no sense to continue.");
-            }
-
-            Slope = new Timeseries();
-            random = new Random();
-            fnn = new BoxAssistedFnn(512, _length);
-            Result = new LyapunovSpectrum(eDim);
+            
+            Result = new double[eDim];
+            Slope = new DataSeries();
+            Log = new StringBuilder();
         }
 
-        public LeSpecSanoSawada(double[] timeSeries) : this(timeSeries, 2, 1, timeSeries.Length, 0, 1.2, 30, false) // 3rd from end parameter (eps) is 0 to obtain further it's default value
+
+        // 3rd from end parameter (eps) is 0 to obtain further it's default value
+        public LeSpecSanoSawada(int eDim) : this(eDim, 1, 0, 0, 1.2, 30, false)
         {
         }
 
-        public LeSpecSanoSawada(double[] timeSeries, int eDim) : this(timeSeries, eDim, 1, timeSeries.Length, 0, 1.2, 30, false) // 3rd from end parameter (eps) is 0 to obtain further it's default value
-        {
-        }
+        public DataSeries Slope { get; set; }
 
-        public LyapunovSpectrum Result { get; }
+        public StringBuilder Log { get; }
 
+        public double[] Result { get; }
+
+        /// <summary>
+        /// Gets method setup info (parameters values)
+        /// </summary>
+        /// <returns></returns>
         public override string ToString() =>
             new StringBuilder()
             .AppendLine("LES by Sano & Sawada")
             .AppendLine($"m = {_eDim}")
             .AppendLine($"τ = {_tau}")
-            .AppendLine($"iterations = {_iterations}")
-            .AppendLine($"min ε = {NumFormat.ToShort(epsmin)}")
-            .AppendLine($"neighbour size increase factor = {NumFormat.ToShort(_epsStep)}")
+            .AppendLine($"iterations = {iterations}")
+            .AppendLine($"min ε = {NumFormatter.ToShort(epsMin)}")
+            .AppendLine($"neighbour size increase factor = {NumFormatter.ToShort(_epsStep)}")
             .AppendLine($"neighbors count = {_minNeighbors}")
             .AppendLine($"invert timeseries = {_inverse}")
             .ToString();
 
-        public override string GetHelp() =>
+        /// <summary>
+        /// Gets help on the method and its params
+        /// </summary>
+        /// <returns></returns>
+        public string GetHelp() =>
             new StringBuilder()
             .AppendLine($"LES by Sano & Sawada [{Paper}]")
             .AppendLine("m - embedding dimension (default: 2)")
@@ -108,27 +108,37 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             .AppendLine($"invert timeseries (default: false)")
             .ToString();
 
-        public override string GetResult() => Result.ToString();
+        public string GetResultAsString() =>
+            new StringBuilder().Append("LE Spectrum: ")
+                .AppendLine(string.Join("; ", Result.Select(le => NumFormatter.ToShort(le))))
+                .AppendLine($"Dky = {NumFormatter.ToShort(StochasticProperties.KYDimension(Result))}")
+                .AppendLine($"Eks = {NumFormatter.ToShort(StochasticProperties.KSEntropy(Result))}")
+                .AppendLine($"PVC = {NumFormatter.ToShort(StochasticProperties.PhaseVolumeContractionSpeed(Result))}")
+                .ToString();
 
-        public override void Calculate()
+        public void Calculate(double[] timeSeries)
         {
-            double[,] delta;
-            double[] dynamics;
-            double[] lfactor;
-            double[] factor;
-            double maxinterval = 0d;
-            int start, i, j;
+            double[] series = new double[timeSeries.Length];
+            Array.Copy(timeSeries, series, series.Length);
 
-            averr = 0.0;
-
-            var interval = Ext.RescaleData(Series);
-
-            if (interval > maxinterval)
+            if (iterations == 0)
             {
-                maxinterval = interval;
+                iterations = series.Length;
             }
 
-            double variance = Ext.Variance(Series);
+            if (_minNeighbors > (series.Length - _tau * (_eDim - 1) - 1))
+            {
+                throw new ArgumentException($"Too few points to find {_minNeighbors} neighbors, it makes no sense to continue.");
+            }
+
+            BoxAssistedFnn fnn = new BoxAssistedFnn(512, series.Length);
+
+            int start, i, j;
+
+            errorAvg = 0.0;
+
+            double interval = Arrays.Rescale(series);
+            double variance = Statistics.Variance(series);
 
             if (variance == 0.0)
             {
@@ -137,24 +147,21 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
             if (_inverse)
             {
-                Array.Reverse(Series);
+                Array.Reverse(series);
             }
 
-            epsmin = 
-                _epsset ? 
-                epsmin / maxinterval : 
-                interval / 1e-3;
+            epsMin = _epsSet ? epsMin / interval : interval / 1e-3;
 
-            dynamics = new double[_eDim];
-            factor = new double[_eDim];
-            lfactor = new double[_eDim];
-            delta = new double[_eDim, _eDim];
+            double[] dynamics = new double[_eDim];
+            double[] factor = new double[_eDim];
+            double[] lfactor = new double[_eDim];
+            double[,] delta = new double[_eDim, _eDim];
             vector = new double[_eDim + 1];
             matrix = new double[_eDim + 1, _eDim + 1];
 
             indexes = MakeIndex(_eDim, _tau);
 
-            random = new Random(int.MaxValue);
+            Random random = new Random(int.MaxValue);
 
             for (i = 0; i < 10000; i++)
             {
@@ -173,16 +180,16 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
             GramSchmidt(delta, lfactor);
 
-            start = Math.Min(_iterations, _length - _tau);
+            start = Math.Min(iterations, series.Length - _tau);
 
-            abstand = new double[_length];
+            abstand = new double[series.Length];
 
             var timer = Stopwatch.StartNew();
 
             for (i = (_eDim - 1) * _tau; i < start; i++)
             {
                 count++;
-                MakeDynamics(dynamics, i);
+                MakeDynamics(series, fnn, dynamics, i);
                 MakeIteration(dynamics, delta);
                 GramSchmidt(delta, lfactor);
 
@@ -199,8 +206,8 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
                     for (j = 0; j < _eDim; j++)
                     {
-                        Log.Append($"{NumFormat.ToShort(factor[j] / count)} ");
-                        Result.Spectrum[j] = factor[j] / count;
+                        Log.Append($"{NumFormatter.ToShort(factor[j] / count)} ");
+                        Result[j] = factor[j] / count;
                     }
 
                     Log.AppendLine();
@@ -208,13 +215,13 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             }
 
             Log.AppendLine();
-            Log.AppendLine($"Avg. rel. forecast error = {NumFormat.ToShort(Math.Sqrt(averr / count) / variance)}");
-            Log.AppendLine($"Avg. abs. forecast error = {NumFormat.ToShort(Math.Sqrt(averr / count) * interval)}");
-            Log.AppendLine($"Avg. neighborhood size   = {NumFormat.ToShort(aveps * maxinterval / count)}");
-            Log.AppendLine($"Avg. number of neighbors = {NumFormat.ToShort(avneig / count)}");
+            Log.AppendLine($"Avg. rel. forecast error = {NumFormatter.ToShort(Math.Sqrt(errorAvg / count) / variance)}");
+            Log.AppendLine($"Avg. abs. forecast error = {NumFormatter.ToShort(Math.Sqrt(errorAvg / count) * interval)}");
+            Log.AppendLine($"Avg. neighborhood size   = {NumFormatter.ToShort(epsAvg * interval / count)}");
+            Log.AppendLine($"Avg. number of neighbors = {NumFormatter.ToShort(neighborsAvg / count)}");
         }
 
-        private double Sort(long act, long nFound, out long nfound, out bool enough)
+        private double Sort(double[] series, BoxAssistedFnn fnn, long act, long nFound, out long nfound, out bool enough)
         {
             double maxeps, dx, dswap, maxdx;
             int self = 0, i, j, del, hf, iswap;
@@ -228,12 +235,12 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
                 if (hf != act)
                 {
-                    maxdx = Math.Abs(Series[act] - Series[hf]);
+                    maxdx = Math.Abs(series[act] - series[hf]);
 
                     for (j = 1; j < _eDim; j++)
                     {
                         del = indexes[j];
-                        dx = Math.Abs(Series[act - del] - Series[hf - del]);
+                        dx = Math.Abs(series[act - del] - series[hf - del]);
                         if (dx > maxdx) maxdx = dx;
                     }
 
@@ -267,7 +274,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 }
             }
 
-            if (!_epsset || (abstand[_minNeighbors - 1] >= epsmin))
+            if (!_epsSet || (abstand[_minNeighbors - 1] >= epsMin))
             {
                 nfound = _minNeighbors;
                 enough = true;
@@ -290,7 +297,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                         fnn.Found[j] = iswap;
                     }
                 }
-                if (abstand[i] > epsmin)
+                if (abstand[i] > epsMin)
                 {
                     nfound = i + 1;
                     enough = true;
@@ -305,15 +312,13 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             return maxeps;
         }
 
-        private void MakeDynamics(double[] dynamics, int act)
+        private void MakeDynamics(double[] series, BoxAssistedFnn fnn, double[] dynamics, int act)
         {
             long i, hi, j, hj, k, t = act;
             long nfound;
-            double[,] imat;
             double foundeps = 0.0, epsilon, hv, hv1;
-            double new_vec;
 
-            epsilon = epsmin / _epsStep;
+            epsilon = epsMin / _epsStep;
 
             do
             {
@@ -324,14 +329,14 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                     epsilon = EpsMax;
                 }
 
-                fnn.PutInBoxes(Series, epsilon, (_eDim - 1) * _tau, _length - _tau, 0, 0);
-                nfound = fnn.FindNeighborsJ(Series, _eDim, _tau, epsilon, act);
+                fnn.PutInBoxes(series, epsilon, (_eDim - 1) * _tau, series.Length - _tau, 0, 0);
+                nfound = fnn.FindNeighborsJ(series, _eDim, _tau, epsilon, act);
 
                 if (nfound > _minNeighbors)
                 {
                     bool got_enough;
 
-                    foundeps = Sort(act, nfound, out nfound, out got_enough);
+                    foundeps = Sort(series, fnn, act, nfound, out nfound, out got_enough);
 
                     if (got_enough)
                     {
@@ -341,12 +346,12 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             } 
             while (epsilon < EpsMax);
 
-            avneig += nfound;
-            aveps += foundeps;
+            neighborsAvg += nfound;
+            epsAvg += foundeps;
 
-            if (!_epsset)
+            if (!_epsSet)
             {
-                epsmin = aveps / count;
+                epsMin = epsAvg / count;
             }
 
             if (nfound < _minNeighbors)
@@ -354,8 +359,8 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 throw new CalculationException("Not enough neighbors found.");
             }
 
-            Ext.FillVectorWith(vector, 0d);
-            Ext.FillMatrixWith(matrix, 0d);
+            Arrays.FillArrayWith(vector, 0d);
+            Matrixes.FillWith(matrix, 0d);
 
             for (i = 0; i < nfound; i++)
             {
@@ -364,17 +369,17 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
                 for (j = 0; j < _eDim; j++)
                 {
-                    matrix[0, j + 1] += Series[act - indexes[j]];
+                    matrix[0, j + 1] += series[act - indexes[j]];
                 }
 
                 for (j = 0; j < _eDim; j++)
                 {
-                    hv1 = Series[act - indexes[j]];
+                    hv1 = series[act - indexes[j]];
                     hj = j + 1;
 
                     for (k = j; k < _eDim; k++)
                     {
-                        matrix[hj, k + 1] += Series[act - indexes[k]] * hv1;
+                        matrix[hj, k + 1] += series[act - indexes[k]] * hv1;
                     }
                 }
             }
@@ -387,19 +392,19 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 }
             }
 
-            imat = InvertMatrix(matrix, _eDim + 1);
+            double[,] imat = InvertMatrix(matrix, _eDim + 1);
 
-            Ext.FillVectorWith(vector, 0d);
+            Arrays.FillArrayWith(vector, 0d);
 
             for (i = 0; i < nfound; i++)
             {
                 act = fnn.Found[i];
-                hv = Series[act + _tau];
+                hv = series[act + _tau];
                 vector[0] += hv;
 
                 for (j = 0; j < _eDim; j++)
                 {
-                    vector[j + 1] += hv * Series[act - indexes[j]];
+                    vector[j + 1] += hv * series[act - indexes[j]];
                 }
             }
 
@@ -408,7 +413,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 vector[i] /= (double)nfound;
             }
 
-            new_vec = 0.0;
+            double new_vec = 0.0;
 
             for (i = 0; i <= _eDim; i++)
             {
@@ -428,18 +433,19 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
             for (i = 0; i < _eDim; i++)
             {
-                new_vec += dynamics[i] * Series[t - indexes[i]];
+                new_vec += dynamics[i] * series[t - indexes[i]];
             }
 
-            averr += (new_vec - Series[t + _tau]) * (new_vec - Series[t + _tau]);
+            double tmp = new_vec - series[t + _tau];
+            errorAvg += tmp * tmp;
 
         }
 
         private void GramSchmidt(double[,] delta, double[] stretch)
         {
             double[,] dnew = new double[_eDim, _eDim];
-            double norm;
             double[] diff = new double[_eDim];
+            double norm;
             long i, j, k;
 
             for (i = 0; i < _eDim; i++)
@@ -491,10 +497,8 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
         private void MakeIteration(double[] dynamics, double[,] delta)
         {
-            double[,] dnew;
             long i, j, k;
-
-            dnew = new double[_eDim, _eDim];
+            double[,] dnew = new double[_eDim, _eDim];
 
             for (i = 0; i < _eDim; i++)
             {
@@ -523,18 +527,14 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
         private double[,] InvertMatrix(double[,] mat, int size)
         {
             int i, j, k;
-            double[,] imat;
             double[] vec = new double[size];
-            double[][] hmat;
-
-            hmat = new double[size][];
+            double[,] imat = new double[size, size];
+            double[][] hmat = new double[size][];
 
             for (int ii = 0; ii < size; ii++)
             {
                 hmat[ii] = new double[size];
             }
-
-            imat = new double[size, size];
 
             for (i = 0; i < size; i++)
             {
@@ -628,7 +628,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             }
         }
 
-        private int[] MakeIndex(int eDim, int delay)
+        private static int[] MakeIndex(int eDim, int delay)
         {
             int[] mmi = new int[eDim];
 
