@@ -4,37 +4,35 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Text;
 using ChaosSoft.Core.IO;
-using ChaosSoft.Core.NumericalMethods.EmbeddingDimension;
+using ChaosSoft.Core.NumericalMethods.PhaseSpace;
+using ChaosSoft.Core.Extensions;
 
 namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 {
     /// <summary>
     /// H. Kantz, A robust method to estimate the maximal Lyapunov exponent of a time series, Phys. Lett. A 185, 77 (1994)
     /// </summary>
-    public class LleKantz : LyapunovMethod
+    public sealed class LleKantz : ITimeSeriesLyapunov, IDescribable
     {
         private const string Paper = "H. Kantz, A robust method to estimate the maximal Lyapunov exponent of a time series, Phys. Lett. A 185, 77 (1994)";
-        private readonly BoxAssistedFnn _fnn;
         private readonly int _eDim;
         private readonly int _tau;
         private readonly int _iterations;
         private readonly int _window; // (0)
         private readonly double _epsMin;
         private readonly double _epsMax;
-        private readonly int _length;
 
-        private int epscount;   // (5)
-        private double epsmin;  //1e-3  
-        private double epsmax;  //1e-2
+        private int epsCount;
+        private double epsMin;
+        private double epsMax;
 
-        private double[] lyap;
-        private int[] count;
+        private readonly double[] _lyap;
+        private readonly int[] _count;
         private int nf;
 
         /// <summary>
         /// The method estimates the largest Lyapunov exponent of a given scalar data set using the algorithm of Kantz.
         /// </summary>
-        /// <param name="series">timeseries to analyze</param>
         /// <param name="eDim">embedding dimension</param>
         /// <param name="tau"></param>
         /// <param name="iterations"></param>
@@ -42,30 +40,39 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
         /// <param name="epsMin"></param>
         /// <param name="epsMax"></param>
         /// <param name="epsCount">number of length scales to use</param>
-        public LleKantz(double[] series, int eDim, int tau, int iterations, int window, double epsMin, double epsMax, int epsCount)
-            : base(series)
+        public LleKantz(int eDim, int tau, int iterations, int window, double epsMin, double epsMax, int epsCount)
         {
             _eDim = eDim;
-
             _tau = tau;
             _iterations = iterations;
             _window = window;
             _epsMin = epsMin;
             _epsMax = epsMax;
-            _length = series.Length;
-            this.epscount = epsCount;
+            this.epsCount = epsCount;
 
-            if (iterations + (eDim - 1) * tau >= _length)
-            {
-                throw new ArgumentException("Too few points to handle specified parameters, it makes no sense to continue.");
-            }
+            _lyap = new double[_iterations + 1];
+            _count = new int[_iterations + 1];
 
-            _fnn = new BoxAssistedFnn(128, _length);
-            SlopesList = new Dictionary<string, Timeseries>();
+            SlopesList = new Dictionary<string, DataSeries>();
+
+            Slope = new DataSeries();
+            Log = new StringBuilder();
         }
 
-        public Dictionary<string, Timeseries> SlopesList { get; set; }
+        public LleKantz(int eDim) : this(eDim, 1, 50, 0, 0, 0, 5)
+        {
+        }
 
+        public DataSeries Slope { get; set; }
+
+        public StringBuilder Log { get; }
+
+        public Dictionary<string, DataSeries> SlopesList { get; set; }
+
+        /// <summary>
+        /// Gets method setup info (parameters values)
+        /// </summary>
+        /// <returns></returns>
         public override string ToString() =>
             new StringBuilder()
             .AppendLine("LLE by Kantz")
@@ -73,11 +80,15 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             .AppendLine($"τ = {_tau}")
             .AppendLine($"iterations = {_iterations}")
             .AppendLine($"theiler window = {_window}")
-            .AppendLine($"min ε = {NumFormat.ToShort(epsmin)}")
-            .AppendLine($"max ε = {NumFormat.ToShort(epsmax)}")
+            .AppendLine($"min ε = {NumFormatter.ToShort(epsMin)}")
+            .AppendLine($"max ε = {NumFormatter.ToShort(epsMax)}")
             .ToString();
 
-        public override string GetHelp() =>
+        /// <summary>
+        /// Gets help on the method and its params
+        /// </summary>
+        /// <returns></returns>
+        public string GetHelp() =>
             new StringBuilder()
             .AppendLine($"LLE by Kantz [{Paper}]")
             .AppendLine("m - embedding dimension (default: 2)")
@@ -88,67 +99,72 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             .AppendLine($"max ε - Max scale (default: 1e-2)")
             .ToString();
 
-        public override string GetResult() => "Successful";
+        public string GetResultAsString() => "Successful";
 
-        public override void Calculate()
+        public void Calculate(double[] timeSeries)
         {
-            double eps_fak;
+            double[] series = new double[timeSeries.Length];
+            Array.Copy(timeSeries, series, series.Length);
+
+            if (_iterations + (_eDim - 1) * _tau >= series.Length)
+            {
+                throw new ArgumentException("Too few points to handle specified parameters, it makes no sense to continue.");
+            }
+
+            BoxAssistedFnn _fnn = new BoxAssistedFnn(128, series.Length);
+
+            double epsFak;
             double epsilon;
             int j,l;
-            var blength = _length - (_eDim - 1) * _tau - _iterations;
+            var blength = series.Length - (_eDim - 1) * _tau - _iterations;
 
-            var interval = Ext.RescaleData(Series);
+            double interval = Arrays.Rescale(series);
 
-            epsmin = 
-                _epsMin == 0 ? 
-                1e-3 : 
-                _epsMin / interval;
+            epsMin = _epsMin == 0 ? 1e-3 : _epsMin / interval;
 
-            epsmax = 
-                _epsMax == 0 ? 
-                1e-2 : 
-                _epsMax / interval;
+            epsMax = _epsMax == 0 ? 1e-2 : _epsMax / interval;
 
-            if (epsmin >= epsmax)
+            if (epsMin >= epsMax)
             {
                 throw new ArgumentException("EpsMin > EpsMax");
             }
 
-            this.epscount = epsmin == epsmax ? 1 : epscount;
+            if (epsMin == epsMax)
+            {
+                epsCount = 1;
+            }
 
             var reference = Math.Min(int.MaxValue, blength);
 
             nf = 0;
-            count = new int[_iterations + 1];
-            lyap = new double[_iterations + 1];
 
-            eps_fak = epscount == 1 ? 1d : Math.Pow(epsmax / epsmin, 1d / (epscount - 1));
+            epsFak = epsCount == 1 ? 1d : Math.Pow(epsMax / epsMin, 1d / (epsCount - 1));
 
-            for (l = 0; l < epscount; l++)
+            for (l = 0; l < epsCount; l++)
             {
-                epsilon = epsmin * Math.Pow(eps_fak, l);
+                epsilon = epsMin * Math.Pow(epsFak, l);
 
-                Array.Clear(count, 0, count.Length);
-                Array.Clear(lyap, 0, lyap.Length);
+                Array.Clear(_count, 0, _count.Length);
+                Array.Clear(_lyap, 0, _lyap.Length);
 
-                _fnn.PutInBoxes(Series, epsilon, 0, blength, 0, _tau);
+                _fnn.PutInBoxes(series, epsilon, 0, blength, 0, _tau);
 
                 for (int i = 0; i < reference; i++)
                 {
-                    nf = _fnn.FindNeighborsK(Series, _eDim, _tau, epsilon, i, _window);
-                    Iterate(i);
+                    nf = _fnn.FindNeighborsK(series, _eDim, _tau, epsilon, i, _window);
+                    Iterate(series, i, _fnn.Found);
                 }
 
                 Log.AppendFormat(CultureInfo.InvariantCulture, "epsilon= {0:F5}\n", epsilon * interval);
 
-                Timeseries dict = new Timeseries();
+                DataSeries dict = new DataSeries();
 
                 for (j = 0; j <= _iterations; j++)
                 {
-                    if (count[j] != 0)
+                    if (_count[j] != 0)
                     {
-                        Log.AppendFormat(CultureInfo.InvariantCulture, "{0}\t{1:F5}\t{2}\n", j, lyap[j] / count[j], count[j]);
-                        dict.AddDataPoint(j, lyap[j] / count[j]);
+                        Log.AppendFormat(CultureInfo.InvariantCulture, "{0}\t{1:F5}\t{2}\n", j, _lyap[j] / _count[j], _count[j]);
+                        dict.AddDataPoint(j, _lyap[j] / _count[j]);
                     }
                 }
                     
@@ -169,7 +185,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             }
         }
 
-        private void Iterate(long act)
+        private void Iterate(double[] series, long act, int[] found)
         {
             double[] lfactor = new double[_iterations + 1];
             double[] dx = new double[_iterations + 1];
@@ -179,11 +195,11 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
               
             for (k = 0; k < nf; k++)
             {
-                element = _fnn.Found[k];
+                element = found[k];
             
                 for (i = 0; i <= _iterations; i++)
                 {
-                    dx[i] = FastMath.Pow2(Series[act + i] - Series[element + i]);
+                    dx[i] = FastMath.Pow2(series[act + i] - series[element + i]);
                 }
 
                 for (l = 1; l < _eDim; l++)
@@ -192,7 +208,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             
                     for (i = 0; i <= _iterations; i++)
                     {
-                        dx[i] += FastMath.Pow2(Series[act + i + l1] - Series[element + l1 + i]);
+                        dx[i] += FastMath.Pow2(series[act + i + l1] - series[element + l1 + i]);
                     }
                 }
             
@@ -210,8 +226,8 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             {
                 if (lcount[j] != 0)
                 {
-                    count[j]++;
-                    lyap[j] += Math.Log(lfactor[j] / lcount[j]) / 2.0;
+                    _count[j]++;
+                    _lyap[j] += Math.Log(lfactor[j] / lcount[j]) / 2.0;
                 }
             }
         }
