@@ -1,4 +1,5 @@
-﻿using ChaosSoft.Core.IO;
+﻿using ChaosSoft.Core.Data;
+using ChaosSoft.Core.IO;
 using System;
 using System.Text;
 
@@ -8,7 +9,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
     /// LLE by Wolf (fixed evolution time)
     /// A Wolf, JB Swift, HL Swinney, JA Vastano, Determining Lyapunov exponents from a time series, Physica D: Nonlinear Phenomena, 1985
     /// </summary>
-    public class LleWolf : LyapunovMethod
+    public sealed class LleWolf : ITimeSeriesLyapunov, IDescribable
     {
         private const string Paper = "A Wolf, JB Swift, HL Swinney, JA Vastano, Determining Lyapunov exponents from a time series, Physica D: Nonlinear Phenomena, 1985";
         
@@ -19,17 +20,19 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
         private readonly double _epsMin;
         private readonly double _epsMax;
 
-        // Pre-calculated variables
-        private readonly double evMulStMulLog2;
-        private readonly int _tsLength;
-
         private readonly double[] _pt1;
-        private readonly double[] _pt2; 
+        private readonly double[] _pt2;
+        
+        // Pre-calculated variables
+        private readonly double _evMulStMulLog2;
 
-        private int propagation;
+        // ---------------------------------
+        private int ind2;
+        private double dii;
+        private double zmult;
+        private double angleMax;
 
-        public LleWolf(double[] series, int eDim, int tau, double dt, double epsMin, double epsMax, int evolv) : 
-            base(series)
+        public LleWolf(int eDim, int tau, double dt, double epsMin, double epsMax, int evolv) : base()
         {
             _eDim = eDim;
             _tau = tau;
@@ -41,34 +44,46 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             _pt1 = new double[_eDim];
             _pt2 = new double[_eDim];
 
-            evMulStMulLog2 = _evolv * _dt * Math.Log(2d);
-            _tsLength = series.Length;
+            _evMulStMulLog2 = _evolv * _dt * Math.Log(2d);
+
+            Slope = new DataSeries();
+            Log = new StringBuilder();
         }
 
-        public LleWolf(double[] timeSeries, int eDim, double stepSize) :
-            this(timeSeries, eDim, 1, stepSize, 1e-3, 1e-2, 1)
+        public LleWolf(int eDim, double stepSize) : this(eDim, 1, stepSize, 1e-3, 1e-2, 1)
         {
         }
 
-        public LleWolf(double[] timeSeries) : 
-            this(timeSeries, 2, 1, 1d, 1e-3, 1e-2, 1)
+        public LleWolf(int eDim) : this(eDim, 1, 1d, 1e-3, 1e-2, 1)
         {
         }
+
+        public DataSeries Slope { get; set; }
+
+        public StringBuilder Log { get; }
 
         public double Result { get; protected set; }
 
+        /// <summary>
+        /// Gets method setup info (parameters values)
+        /// </summary>
+        /// <returns></returns>
         public override string ToString() =>
             new StringBuilder()
                 .AppendLine("LLE by Wolf (fixed evolution time)")
                 .AppendLine($"m = {_eDim}")
                 .AppendLine($"τ = {_tau}")
-                .AppendLine($"Δt = {NumFormat.ToShort(_dt)}")
-                .AppendLine($"min ε = {NumFormat.ToShort(_epsMin)}")
-                .AppendLine($"max ε = {NumFormat.ToShort(_epsMax)}")
+                .AppendLine($"Δt = {NumFormatter.ToShort(_dt)}")
+                .AppendLine($"min ε = {NumFormatter.ToShort(_epsMin)}")
+                .AppendLine($"max ε = {NumFormatter.ToShort(_epsMax)}")
                 .AppendLine($"evolution steps: {_evolv}")
                 .ToString();
 
-        public override string GetHelp() =>
+        /// <summary>
+        /// Gets help on the method and its params
+        /// </summary>
+        /// <returns></returns>
+        public string GetHelp() =>
             new StringBuilder()
                 .AppendLine($"LLE by Wolf (fixed evolution time) [{Paper}]")
                 .AppendLine("m - embedding dimension (default: 2)")
@@ -79,33 +94,38 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 .AppendLine("evolution steps (default: 1)")
                 .ToString();
 
-        public override string GetResult() => NumFormat.ToShort(Result);
+        /// <summary>
+        /// Gets result in string format
+        /// </summary>
+        /// <returns></returns>
+        public string GetResultAsString() => NumFormatter.ToShort(Result);
 
-        public override void Calculate()
+        public void Calculate(double[] timeSeries)
         {
             double zlyap = 0;
-            ////int ind = 0; // <In fortran was 1>.
-            int ind2 = 0; // Points to second trajectory <in fortran was empty and defined later>.
+            ////int ind = 0; // In fortran was 1
             double sum = 0d; // Holds running exponent estimate sans i/time.
             int its = 0; // Total number of propagation steps.
-
-            double dii = 0; // <initialization absent in fortran>
-
-            // Calculate useful size of timeseries.
-            int dataPointsCount = _tsLength - _eDim * _tau - _evolv;
+            int propagation;
 
             // Find nearest neighbor to first data point.
             double di = 1e38;
+
+            // Calculate useful size of timeseries.
+            int dataPointsCount = timeSeries.Length - _eDim * _tau - _evolv;
+
+            dii = 0; // <initialization absent in fortran>
+            ind2 = 0; // Points to second trajectory <in fortran was empty and defined later>.
 
             // Don't take point too close to fiducial point.
             for (int i = 10; i < dataPointsCount; i++)
             {
                 // Compute separation between fiducial point and candidate.
-                var d = 0d;
+                double d = 0d;
 
                 for (int j = 0; j < _eDim; j++)
                 {
-                    d += FastMath.Pow2(GetAttractorPoint(0, j) - GetAttractorPoint(i, j));
+                    d += FastMath.Pow2(GetAttractorPoint(timeSeries, 0, j) - GetAttractorPoint(timeSeries, i, j));
                 }
 
                 d = Math.Sqrt(d);
@@ -126,8 +146,8 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 // Get coordinates of evolved points.
                 for (int j = 0; j < _eDim; j++)
                 {
-                    _pt1[j] = GetAttractorPoint(ind, j);
-                    _pt2[j] = GetAttractorPoint(ind2 + _evolv, j);
+                    _pt1[j] = GetAttractorPoint(timeSeries, ind, j);
+                    _pt2[j] = GetAttractorPoint(timeSeries, ind2 + _evolv, j);
                 }
 
                 // Compute final separation between pair, update exponent.
@@ -141,7 +161,7 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
                 df = Math.Sqrt(df);
 
                 its++;
-                sum += Math.Log(df / di) / evMulStMulLog2;
+                sum += Math.Log(df / di) / _evMulStMulLog2;
                 zlyap = sum / its;
 
                 propagation = _evolv * its;
@@ -150,85 +170,16 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
 
                 // Look for replacement point.
                 int indold = ind2;
-                double zmult = 1d; // Multiplier of scalMax when go to longer distances.
-                double anglmx = 0.3;
+                zmult = 1d; // Multiplier of scalMax when go to longer distances.
+                angleMax = 0.3;
 
-                metka:
+                bool calculated;
 
-                double thmin = Math.PI;
-
-                // Search over all points.
-                for (int i = 0; i < dataPointsCount; i++)
+                do
                 {
-                    // Don't take points too close in time to fiducial point.
-                    int iii = Math.Abs(i - ind);
-
-                    if (iii < 9)
-                    {
-                        continue;
-                    }
-
-                    // Compute distance between fiducial point and candidate.
-                    double dnew = 0d;
-
-                    for (int j = 0; j < _eDim; j++)
-                    {
-                        dnew += FastMath.Pow2(_pt1[j] - GetAttractorPoint(i, j));
-                    }
-
-                    dnew = Math.Sqrt(dnew);
-
-                    // Look further away than noise scale, closer than zmult*scalMax.
-                    if (dnew > zmult * _epsMax || dnew < _epsMin)
-                    {
-                        continue;
-                    }
-
-                    // Find angular change old to new vector.
-                    double dot = 0d;
-
-                    for (int j = 0; j < _eDim; j++)
-                    {
-                        dot += (_pt1[j] - GetAttractorPoint(i, j)) * (_pt1[j] - _pt2[j]);
-                    }
-
-                    double cth = Math.Abs(dot / (dnew * df));
-
-                    cth = FastMath.Min(cth, 1d);
-
-                    double th = Math.Acos(cth);
-
-                    // Save point with smallest angular change so far.
-                    if (th < thmin)
-                    { 
-                        thmin = th;
-                        dii = dnew;
-                        ind2 = i;
-                    }
+                    calculated = LookForReplacementPoint(timeSeries, dataPointsCount, ind, indold, df);
                 }
-
-                if (thmin > anglmx)
-                {
-                    // Can't find a replacement - look at longer distances.
-                    zmult++;
-
-                    if (zmult <= 5d)
-                    {
-                        goto metka;
-                    }
-
-                    // No replacement at 5*scale, double search angle, reset distance.
-                    zmult = 1d;
-                    anglmx *= 2d;
-
-                    if (anglmx < Math.PI)
-                    {
-                        goto metka;
-                    }
-
-                    ind2 = indold + _evolv;
-                    dii = df;
-                }
+                while (!calculated);
 
                 di = dii;
             }
@@ -236,12 +187,92 @@ namespace ChaosSoft.Core.NumericalMethods.Lyapunov
             Result = zlyap;
         }
 
+        private bool LookForReplacementPoint(double[] series, int dataPointsCount, int ind, int indold, double df)
+        {
+            double thmin = Math.PI;
+
+            // Search over all points.
+            for (int i = 0; i < dataPointsCount; i++)
+            {
+                // Don't take points too close in time to fiducial point.
+                int iii = Math.Abs(i - ind);
+
+                if (iii < 9)
+                {
+                    continue;
+                }
+
+                // Compute distance between fiducial point and candidate.
+                double dnew = 0d;
+
+                for (int j = 0; j < _eDim; j++)
+                {
+                    dnew += FastMath.Pow2(_pt1[j] - GetAttractorPoint(series, i, j));
+                }
+
+                dnew = Math.Sqrt(dnew);
+
+                // Look further away than noise scale, closer than zmult*scalMax.
+                if (dnew > zmult * _epsMax || dnew < _epsMin)
+                {
+                    continue;
+                }
+
+                // Find angular change old to new vector.
+                double dot = 0d;
+
+                for (int j = 0; j < _eDim; j++)
+                {
+                    dot += (_pt1[j] - GetAttractorPoint(series, i, j)) * (_pt1[j] - _pt2[j]);
+                }
+
+                double cth = Math.Abs(dot / (dnew * df));
+
+                cth = FastMath.Min(cth, 1d);
+
+                double th = Math.Acos(cth);
+
+                // Save point with smallest angular change so far.
+                if (th < thmin)
+                {
+                    thmin = th;
+                    dii = dnew;
+                    ind2 = i;
+                }
+            }
+
+            if (thmin > angleMax)
+            {
+                // Can't find a replacement - look at longer distances.
+                zmult++;
+
+                if (zmult <= 5d)
+                {
+                    return false;
+                }
+
+                // No replacement at 5*scale, double search angle, reset distance.
+                zmult = 1d;
+                angleMax *= 2d;
+
+                if (angleMax < Math.PI)
+                {
+                    return false;
+                }
+
+                ind2 = indold + _evolv;
+                dii = df;
+            }
+
+            return true;
+        }
+
         ///<summary>Define delay coordinates with a statement function</summary>
         ///<returns>j-th component of i-th reconstructed attractor point</returns>
-        private double GetAttractorPoint(int i, int j)
+        private double GetAttractorPoint(double[] series, int i, int j)
         {
             int point = i + j * _tau;
-            return point < _tsLength ? Series[point] : 0;
+            return point < series.Length ? series[point] : 0;
         }
     }
 }
